Remove `net` module.
Useful types were moved to the `support` module.
