+++ rsomeip/Cargo.toml
@@ -9,7 +9,6 @@ description = "A Rust implementation of AUTOSTAR's Scalable service-Oriented Mid
 
 [dependencies]
 bytes = "1.8.0"
-thiserror = "1.0.50"
 tokio = { version = "1.29.1", features = [
     "net",
     "rt",
