--- rsomeip/src/net.rs
@@ -1,3 +1,0 @@
-//! Network communication using SOME/IP, TCP and UDP.
-
-pub mod util;
