--- rsomeip/src/net/error/tests.rs
@@ -1,10 +1,0 @@
-use super::*;
-
-#[test]
-fn error_from_str() {
-    let error = Error::Failure("bug");
-    assert_eq!(
-        error.to_string(),
-        String::from("operation failed unexpectedly: bug")
-    );
-}
