+++ rsomeip/src/support.rs
@@ -1,0 +1,7 @@
+//! Support module for the rest of rsomeip.
+//!
+//! Contains utilities that don't fit in anywhere else.
+
+pub mod collections;
+#[allow(unused)]
+pub mod sync;
